# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Padel Backend service.",
        "title": "Padel Backend API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/organizations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all organizations in a specific city",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "List organizations by city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City name",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ListOrganizationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new organization in the city",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "Create a new organization",
                "parameters": [
                    {
                        "description": "Organization payload",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.CreateOrganizationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/organizations/{orgID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves an organization by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "Get an organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.OrganizationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "Update an organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Organization payload",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.UpdateOrganizationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an organization by ID",
                "tags": [
                    "organizations"
                ],
                "summary": "Delete an organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_controllers_http.CreateOrganizationRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "City is the city in which the organization itself is located\nexample: Almaty",
                    "type": "string",
                    "example": "Astana"
                },
                "name": {
                    "description": "Name is a organization name\nexample: Padel Club #1",
                    "type": "string",
                    "example": "Padel club"
                }
            }
        },
        "internal_controllers_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Message provides a human-readable description of the error.",
                    "type": "string",
                    "example": "invalid JSON body"
                }
            }
        },
        "internal_controllers_http.ListOrganizationsResponse": {
            "type": "object",
            "properties": {
                "organizations": {
                    "description": "Organizations is the list of organizations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_controllers_http.OrganizationResponse"
                    }
                }
            }
        },
...
        "internal_controllers_http.OrganizationResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "City is the city where the organization is located\nexample: Almaty",
                    "type": "string",
                    "example": "Almaty"
                },
                "createdAt": {
                    "description": "CreatedAt is the timestamp when the organization was created\nexample: 2025-11-01T10:00:00Z",
                    "type": "string",
                    "example": "2025-11-01T10:00:00Z"
                },
                "id": {
                    "description": "ID is the unique identifier of the created organization\nexample: org-123",
                    "type": "string",
                    "example": "org-1"
                },
                "name": {
                    "description": "Name is the organization name\nexample: Padel Club Almaty",
                    "type": "string",
                    "example": "Padel Club Almaty"
                },
                "updatedAt": {
                    "description": "UpdatedAt is the timestamp when the organization was last updated (RFC3339 format)\nexample: 2025-11-01T10:00:00Z",
                    "type": "string",
                    "example": "2025-11-01T10:00:00Z"
                }
            }
        },
...
        "internal_controllers_http.UpdateOrganizationRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "City is the updated city where the organization is located\nexample: Astana",
                    "type": "string",
                    "example": "Astana"
                },
                "name": {
                    "description": "Name is the updated organization name\nexample: Updated Padel Club",
                    "type": "string",
                    "example": "Updated Padel Club"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}